requirements.txt
Flask
flasgger
firebase-admin
APScheduler
gunicorn
requests
python-dotenv
joblib
numpy
flask_cors
scikit-learn
xgboost